% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_D_final_plan.R
\name{.mix_D_final_plan}
\alias{.mix_D_final_plan}
\title{Transform the pseudo-mixture design into a mixture design}
\usage{
.mix_D_final_plan(
  data = NULL,
  mix_plan = NULL,
  lc_Ingredient1 = NULL,
  uc_Ingredient1 = NULL,
  lc_Ingredient2 = NULL,
  uc_Ingredient2 = NULL,
  lc_Ingredient3 = NULL,
  uc_Ingredient3 = NULL
)
}
\arguments{
\item{data}{The pseudo-mixture design from .pseudo.mixture function.}

\item{mix_plan}{The mixture design from mix_design function.}

\item{lc_Ingredient1}{An integer giving the lower constraint on ingredient 1.}

\item{uc_Ingredient1}{An integer giving the upper constraint on ingredient 1.}

\item{lc_Ingredient2}{An integer giving the lower constraint on ingredient 2.}

\item{uc_Ingredient2}{An integer giving the upper constraint on ingredient 2.}

\item{lc_Ingredient3}{An integer giving the lower constraint on ingredient 3.}

\item{uc_Ingredient3}{An integer giving the upper constraint on ingredient 3.}
}
\value{
A data frame.
}
\description{
Internal function.
The function transforms the pseudo-mixture design into a mixture design.
}
\examples{
\dontrun{
.mix_D_final_plan(Dinter[.x, ],
                  mix_plan = mixture,
                  lc_Ingredient1 = lc.Ingredient1, uc_Ingredient1 = uc.Ingredient1,
                  lc_Ingredient2 = lc.Ingredient2, uc_Ingredient2 = uc.Ingredient2,
                  lc_Ingredient3 = lc.Ingredient3, uc_Ingredient3 = uc.Ingredient3)
}
}
\keyword{internal}
