% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_design.R
\name{.mix_design}
\alias{.mix_design}
\title{Build an extreme vertices design in a constrained mixture space}
\usage{
.mix_design(
  Total = NULL,
  lc_Ingredient1 = NULL,
  uc_Ingredient1 = NULL,
  lc_Ingredient2 = NULL,
  uc_Ingredient2 = NULL,
  lc_Ingredient3 = NULL,
  uc_Ingredient3 = NULL
)
}
\arguments{
\item{Total}{An integer giving the constant total amount of the sum of mixture components in the design.}

\item{lc_Ingredient1}{An integer giving the lower constraint on ingredient 1.}

\item{uc_Ingredient1}{An integer giving the upper constraint on ingredient 1.}

\item{lc_Ingredient2}{An integer giving the lower constraint on ingredient 2.}

\item{uc_Ingredient2}{An integer giving the upper constraint on ingredient 2.}

\item{lc_Ingredient3}{An integer giving the lower constraint on ingredient 3.}

\item{uc_Ingredient3}{An integer giving the upper constraint on ingredient 3.}
}
\value{
A data frame.
}
\description{
Internal function.
This function behaves just like Xvert() from mixexp.
The main difference is that the result is expressed in content in the complete recipe and not in relative proportion in the mixing plan.
}
\examples{
\dontrun{
.mix_design(Total = Total,
            lc_Ingredient1 = lc.Ingredient1, uc_Ingredient1 = uc.Ingredient1,
            lc_Ingredient2 = lc.Ingredient2, uc_Ingredient2 = uc.Ingredient2,
            lc_Ingredient3 = lc.Ingredient3, uc_Ingredient3 = uc.Ingredient3)
}
}
\keyword{internal}
