% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrF2mix.R
\name{FrF2mix}
\alias{FrF2mix}
\title{Build a fractional factorial design with mixture constraints}
\usage{
FrF2mix(
  nruns = NULL,
  process.factor.names = NULL,
  Total = NULL,
  lc.Ingredient1 = NULL,
  uc.Ingredient1 = NULL,
  lc.Ingredient2 = NULL,
  uc.Ingredient2 = NULL,
  lc.Ingredient3 = NULL,
  uc.Ingredient3 = NULL
)
}
\arguments{
\item{nruns}{A number of runs, must be a power of 2 (4 to 4096), if given.}

\item{process.factor.names}{A list a nfactors factor names.}

\item{Total}{An integer giving the constant total amount of the sum of mixture components in the design.}

\item{lc.Ingredient1}{An integer giving the lower constraint on ingredient 1.}

\item{uc.Ingredient1}{An integer giving the upper constraint on ingredient 1.}

\item{lc.Ingredient2}{An integer giving the lower constraint on ingredient 2.}

\item{uc.Ingredient2}{An integer giving the upper constraint on ingredient 2.}

\item{lc.Ingredient3}{An integer giving the lower constraint on ingredient 3.}

\item{uc.Ingredient3}{An integer giving the upper constraint on ingredient 3.}
}
\value{
A list with the fractional factorial design with mixture constraints and the reminder of the regular fractional factorial 2-levels designs.
}
\description{
The function edits a DoE combining process and mixing factors, easily dealing with any mixture constraints for 3 ingredients within the framework of fractional factorial designs.
}
\examples{
\dontrun{
FrF2mix(nruns = 16,
        process.factor.names = list(time.hydration = c("30s","2min30s"),
                                    speed.hydration = c("V1", "V3"),
                                    time.emulsification = c("1min30s", "3min30s"),
                                    speed.emulsification = c("V1", "V3")
        ),
        Total = 91.24,
        lc.Ingredient1 = 1, uc.Ingredient1 = 5,
        lc.Ingredient2 = 20, uc.Ingredient2 = 75,
        lc.Ingredient3 = 15, uc.Ingredient3 = 75)
}
}
